using __all__ everywhere, consolidating some function into main _func files
