+++ porespy/filters/__init__.py
@@ -54,42 +54,11 @@ image, but with altered values.
     filters.trim_saddle_points
     filters.trim_small_clusters
 
-
 """
 
-from ._funcs import apply_chords
-from ._funcs import apply_chords_3D
-from ._funcs import apply_padded
-from ._funcs import chunked_func
-from ._funcs import distance_transform_lin
-from ._funcs import fill_blind_pores
-from ._funcs import find_disconnected_voxels
-from ._funcs import find_trapped_regions
-from ._funcs import find_dt_artifacts
-from ._funcs import flood
-from ._funcs import flood_func
-from ._funcs import hold_peaks
-from ._funcs import ibip
-from ._funcs import ibip_gpu
-from ._funcs import local_thickness
-from ._funcs import nphase_border
-from ._funcs import porosimetry
-from ._funcs import prune_branches
-from ._funcs import region_size
-from ._funcs import trim_disconnected_blobs
-from ._funcs import trim_extrema
-from ._funcs import trim_floating_solid
-from ._funcs import trim_nonpercolating_paths
-from ._funcs import trim_small_clusters
-from ._snows import snow_partitioning
-from ._snows import snow_partitioning_n
-from ._snows import snow_partitioning_parallel
-from ._snows import find_peaks
-from ._snows import reduce_peaks
-from ._snows import trim_nearby_peaks
-from ._snows import trim_saddle_points
-from ._snows import trim_saddle_points_legacy
+from ._funcs import *
+from ._snows import *
 from ._size_seq_satn import *
-from ._nlmeans import nl_means_layered
-from ._fftmorphology import fftmorphology
+from ._nlmeans import *
+from ._fftmorphology import *
 from . import imagej
