+++ porespy/filters/_funcs.py
@@ -17,6 +17,34 @@ from porespy import settings
 from porespy.tools import get_tqdm
 
 
+__all__ = [
+    "apply_chords",
+    "apply_chords_3D",
+    "apply_padded",
+    "chunked_func",
+    "distance_transform_lin",
+    "fill_blind_pores",
+    "find_disconnected_voxels",
+    "find_trapped_regions",
+    "find_dt_artifacts",
+    "flood",
+    "flood_func",
+    "hold_peaks",
+    "ibip",
+    "ibip_gpu",
+    "local_thickness",
+    "nphase_border",
+    "porosimetry",
+    "prune_branches",
+    "region_size",
+    "trim_disconnected_blobs",
+    "trim_extrema",
+    "trim_floating_solid",
+    "trim_nonpercolating_paths",
+    "trim_small_clusters",
+]
+
+
 tqdm = get_tqdm()
 logger = logging.getLogger(__name__)
