+++ porespy/generators/_noise.py
@@ -3,6 +3,11 @@ from porespy.tools import norm_to_uniform
 import psutil
 
 
+__all__ = [
+    'fractal_noise',
+]
+
+
 def fractal_noise(
     shape,
     frequency=0.05,
@@ -33,17 +38,21 @@ def fractal_noise(
     mode : string
         The type of noise to generate. Options are:
 
-        - 'perlin'
-            classic Perlin noise
-
-        - 'simplex'
-            updated Perlin noise for more realistic textures
-
-        - 'value'
-            bilnear interpolation of white noise
-
-        - 'cubic'
-            cubic interpolation of white noise
+        ========== ==================================================================
+        mode       description
+        ========== ==================================================================
+        'perlin'   Classic Perlin noise. For more information on ``perlin noise`` see
+                   `here <https://en.wikipedia.org/wiki/Perlin_noise>`__.
+        'simplex'  Updated Perlin noise for more realistic textures. For more
+                   information on ``simplex noise`` see
+                   `here <https://en.wikipedia.org/wiki/Simplex_noise>`__.
+        'value'    Bilnear interpolation of white noise. For more information on
+                   ``value noise`` see
+                   `here <https://en.wikipedia.org/wiki/Value_noise>`__.
+        'cubic'    Cubic interpolation of white noise. For more information on
+                   ``cubic noise`` see
+                   `here <https://github.com/jobtalle/CubicNoise>`__.
+        ========== ==================================================================
 
     seed : int, optional
         The seed of the random number generator.  Using the same
@@ -67,18 +76,6 @@ def fractal_noise(
     To access the more elaborate functionality and options of these
     packages, explore the `pyfastnoisesimd documentation
     <https://pyfastnoisesimd.readthedocs.io/en/latest/overview.html>`_.
-
-    For more information on ``simplex noise`` see
-    `here <https://en.wikipedia.org/wiki/Simplex_noise>`__.
-
-    For more information on ``perlin noise`` see
-    `here <https://en.wikipedia.org/wiki/Perlin_noise>`__.
-
-    For more information on ``value noise`` see
-    `here <https://en.wikipedia.org/wiki/Value_noise>`__.
-
-    For more information on ``cubic noise`` see
-    `here <https://github.com/jobtalle/CubicNoise>`__.
 
     Examples
     --------
