+++ porespy/metrics/__init__.py
@@ -40,27 +40,7 @@ but a few functions can be applied directly to the binary image.
 
 """
 
-from ._regionprops import regionprops_3D
-from ._regionprops import props_to_DataFrame
-from ._regionprops import prop_to_image
-from ._funcs import chord_counts
-from ._funcs import chord_length_distribution
-from ._funcs import find_h
-from ._funcs import lineal_path_distribution
-from ._funcs import pore_size_distribution
-from ._funcs import radial_density_distribution
-from ._funcs import porosity
-from ._funcs import porosity_profile
-from ._funcs import representative_elementary_volume
-from ._funcs import satn_profile
-from ._funcs import two_point_correlation
-from ._funcs import phase_fraction
-from ._funcs import pc_curve
-from ._funcs import pc_curve_from_ibip
-from ._funcs import pc_curve_from_mio
+from ._regionprops import *
+from ._funcs import *
 from ._fractal_dims import boxcount
-from ._meshtools import mesh_surface_area
-from ._meshtools import mesh_volume
-from ._meshtools import region_interface_areas
-from ._meshtools import region_surface_areas
-from ._meshtools import region_volumes
+from ._meshtools import *
