+++ porespy/networks/__init__.py
@@ -23,16 +23,9 @@ Contains functions for analysing images as pore networks.
 
 """
 
-from ._funcs import add_boundary_regions
-from ._funcs import generate_voxel_image
-from ._funcs import label_phases
-from ._funcs import label_boundaries
-from ._funcs import map_to_regions
-from ._maximal_ball import maximal_ball_wrapper
-from ._getnet import regions_to_network
-from ._snow2 import snow2
-from ._utils import _net_dict
-from ._snow2 import _parse_pad_width
-from ._size_factors import diffusive_size_factor_AI
-from ._size_factors import create_model
-from ._size_factors import find_conns
+from ._funcs import *
+from ._maximal_ball import *
+from ._getnet import *
+from ._snow2 import *
+from ._utils import *
+from ._size_factors import *
