+++ porespy/networks/_utils.py
@@ -1,2 +1,9 @@
+
+
+__all__ = [
+    "_net_dict",
+]
+
+
 class _net_dict(dict):
     pass
