+++ porespy/tools/__init__.py
@@ -47,8 +47,6 @@ ways that do NOT return a modified version of the original image.
 
 from ._funcs import *
 from ._utils import *
-from ._funcs import _check_for_singleton_axes
-from ._unpad import unpad
 from ._sphere_insertions import *
